avoid dying on "undefined" requests
The problem here is that liballprof does not have enough info to know if a
request given to wait/waitall is mpi_request_null or not, so doing a wait
on a null request will lead to an error but is perfectly legal. Of course
this patch is dangerous, as it might hide problems, i.e., if req_size is
set to 4 but should be 8. There is no way to do the right thing in all cases,
unless we change liballprof.